requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.0

# Scraping
playwright==1.49.1
playwright-stealth==2.0.2
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1
tenacity==9.0.0
apify-client==2.5.0
exa-py==1.10.1

# Sentiment Analysis
vaderSentiment==3.3.2
google-generativeai==0.8.3

# Background Tasks
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
